It Builds
I finished a lot of things, haven't done the sky shader, but that is to
come.  I also need to implement the new shader loading system.

+++ Source/QuicksandEngine/Network/Network.hpp
@@ -329,7 +329,7 @@ class NetworkGameView : public IGameView
 public:
 	// IGameView Implementation - everything is stubbed out.
 	virtual HRESULT VOnRestore() { return S_OK; }
-	virtual void VOnRender(double fTime, float fElapsedTime) { }
+	virtual void VOnRender(double fTime, double fElapsedTime) { }
 	virtual HRESULT VOnLostDevice() { return S_OK; }
 	virtual GameViewType VGetType() { return GameView_Remote; }
 	virtual GameViewId VGetId() const { return m_ViewId; }
